pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-tasks-db-api"
version = "0.3.1"
description = "Django Rest Framework API for django-tasks-db"
requires-python = ">=3.10"
license = "BSD-3-Clause"
dependencies = [
    "requests>=2.28",
]

[project.optional-dependencies]
server = [
    "Django>=4.2",
    "djangorestframework>=3.14",
    "django-tasks-db>=0.12.0",
    "django-filter>=23.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-django",
    "coverage",
]

[project.urls]
Homepage = "https://github.com/player1537-playground/claude-playground"

[tool.setuptools.packages.find]
include = ["django_tasks_db_api*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_tasks_db_api.tests.settings"
pythonpath = ["."]
